CH4/OFI: Reimplement huge message transfer
The huge message transfer protocol had a major design flaw that it used
a struct to track the status of the transfer and stored it a hash where
the key was the destination/source rank (at the sender/receiver
respectively). This meant that if two transfers were started at the same
time, the first one would be lost when the second one overwrote it.

Here, we revamp the protocol to hash based on the handle of the main
request object. This is safe because there will only be one request that
matches a particular tranfer operation and there will be no overlap
between other operations.

This also improves the documentation around some of the functions
involved in the huge message transfer protocol.

Fixes csr/mpich-ofi#26
Fixes csr/mpich-ofi#483

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>