CH4/OFI: Enable Dynamic Process Support.
Adding the management of connection and functionality for dynamic
process. The connection between two process groups is maintained between
two roots. Each root will get a conn_id for any new communication, which
in turn saved in fi_av and mapped to corresponding fi_addr_t. The
mapped address is saved in a connection manager (part of MPIDI_Global).
Note the connection is only used for management purpose, not the actual
communication between groups.

The network addresses of the processes in each group will get exchanged
during the connect/accept procedure. They will be inserted to fi_av and
mapped to fi_addr_t(s). Any address that is not previously known will be
associated with a new CH4 level AVT_ID. This is a part of the existing
protocol for UPIDS to LUPIDS conversion. After the address exchange, all
process will communicate in a "connection-less" manner.

Disconnection protocol. This patch uses a "eventually disconnect"
protocol for handling disconnection. This means no disconnection message
will be send during the invocation of 'MPI_Comm_disconnect'. The actual
disconnection happens in the destroy of connection manager at finalize
time. Within the connection manager, each connection will be marked as
`CONNECTED_PARENT` and `CONNECTED_CHILD`. The "child" process (root of
the connecting group) will send the "disconnect message" to the "parent"
process (root of the accepting group). The message will be sent using
`fi_tsendmsg` with `FI_TRANSMIT_COMPLETE` flag to avoid delivery problem
when `FI_PROGRESS_MANUAL` is used.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>