ch4/generic: add framework for new long message protocol
Add new handlers for the AM messages in PIPELINE protocol.  The
existing LMT protocol is kept and will be removed in a separate
commits when the PIPELINE is implemented.

For the PIPELINE protocol, it has three new AM message types: RTS, CTS
and SEG. The RTS contains information for initiating the pipeline
protocol. The first segment of data will be piggybacked on the
RTS. The CTS is confirmation from receive side that the send has been
matched and ready for data. SEG is the transmission of data segments.