test/mpi: Add communicator refcount test
Add a test that receives a synchronous send *after* the user has freed
the comm in the application. MPI is required to keep the communicator
active because of pending communication. In this case it is needed for
the ACK message to the sender. ch4/am code previously did not pass
this test.