ch4/ucx: Cleanup pack/unpack datatype handling
UCX generic datatypes take a context pointer from the user. MPICH passes a
pointer to the MPIR_Datatype struct, but then casted/dereferenced that as an
MPI_Datatype pointer. This worked because the datatype handle is the first
element in the struct, but is fragile and confusing to the reader. Make it more
explicit by accessing the handle directly.