pmi: Fix handling of initack response in v1 wire protocol
In the v1 wire protocol, the initack response is an exception in that is
spans multiple lines but is not an mcmd. For example:

  cmd=initack rc=0 rank=0 size=2 appnum=0 debug=1
  cmd=set size=2
  cmd=set rank=0
  cmd=set debug=1

Only the first cmd value is relevant, so we need to be careful not to
overwrite it in the PMIU_cmd struct. The size, rank, and debug values
can be retrieved from the same PMIU_cmd struct.