ch4/ipc: Replace mapped gpu ipc handle cache
Remove mapped buffer cache using AVL trees and replace with a hash
table. The key for the hash is the combination of the sender base
address and their local rank on the node. Each entry can store a mapped
address for each device visible to the process doing the mapping.